// src/protos/bchrpc_pb.d.ts
  export type AsObject = {
  }
// ...
export class Transaction extends jspb.Message {
  getHash(): Uint8Array | string;
  getHash_asU8(): Uint8Array;
  getHash_asB64(): string;
  setHash(value: Uint8Array | string): Transaction;

  getVersion(): number;
  setVersion(value: number): Transaction;

  getInputsList(): Array<Transaction.Input>;
  setInputsList(value: Array<Transaction.Input>): Transaction;
  clearInputsList(): Transaction;
  addInputs(value?: Transaction.Input, index?: number): Transaction.Input;

  getOutputsList(): Array<Transaction.Output>;
  setOutputsList(value: Array<Transaction.Output>): Transaction;
  clearOutputsList(): Transaction;
  addOutputs(value?: Transaction.Output, index?: number): Transaction.Output;

  getLockTime(): number;
  setLockTime(value: number): Transaction;

  getSize(): number;
  setSize(value: number): Transaction;

  getTimestamp(): number;
  setTimestamp(value: number): Transaction;

  getConfirmations(): number;
  setConfirmations(value: number): Transaction;

  getBlockHeight(): number;
  setBlockHeight(value: number): Transaction;

  getBlockHash(): Uint8Array | string;
  getBlockHash_asU8(): Uint8Array;
  getBlockHash_asB64(): string;
  setBlockHash(value: Uint8Array | string): Transaction;

  getSlpTransactionInfo(): SlpTransactionInfo | undefined;
  setSlpTransactionInfo(value?: SlpTransactionInfo): Transaction;
  hasSlpTransactionInfo(): boolean;
  clearSlpTransactionInfo(): Transaction;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Transaction.AsObject;
  static toObject(includeInstance: boolean, msg: Transaction): Transaction.AsObject;
  static serializeBinaryToWriter(message: Transaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Transaction;
  static deserializeBinaryFromReader(message: Transaction, reader: jspb.BinaryReader): Transaction;
}

export namespace Transaction {
  export type AsObject = {
    hash: Uint8Array | string,
    version: number,
    inputsList: Array<Transaction.Input.AsObject>,
    outputsList: Array<Transaction.Output.AsObject>,
    lockTime: number,
    size: number,
    timestamp: number,
    confirmations: number,
    blockHeight: number,
    blockHash: Uint8Array | string,
    slpTransactionInfo?: SlpTransactionInfo.AsObject,
  }

  export class Input extends jspb.Message {
    getIndex(): number;
    setIndex(value: number): Input;

    getOutpoint(): Transaction.Input.Outpoint | undefined;
    setOutpoint(value?: Transaction.Input.Outpoint): Input;
    hasOutpoint(): boolean;
    clearOutpoint(): Input;

    getSignatureScript(): Uint8Array | string;
    getSignatureScript_asU8(): Uint8Array;
    getSignatureScript_asB64(): string;
    setSignatureScript(value: Uint8Array | string): Input;

    getSequence(): number;
    setSequence(value: number): Input;

    getValue(): number;
    setValue(value: number): Input;

    getPreviousScript(): Uint8Array | string;
    getPreviousScript_asU8(): Uint8Array;
    getPreviousScript_asB64(): string;
    setPreviousScript(value: Uint8Array | string): Input;

    getAddress(): string;
    setAddress(value: string): Input;

    getSlpToken(): SlpToken | undefined;
    setSlpToken(value?: SlpToken): Input;
    hasSlpToken(): boolean;
    clearSlpToken(): Input;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Input.AsObject;
    static toObject(includeInstance: boolean, msg: Input): Input.AsObject;
    static serializeBinaryToWriter(message: Input, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Input;
    static deserializeBinaryFromReader(message: Input, reader: jspb.BinaryReader): Input;
  }

  export namespace Input {
    export type AsObject = {
      index: number,
      outpoint?: Transaction.Input.Outpoint.AsObject,
      signatureScript: Uint8Array | string,
      sequence: number,
      value: number,
      previousScript: Uint8Array | string,
      address: string,
      slpToken?: SlpToken.AsObject,
    }

    export class Outpoint extends jspb.Message {
      getHash(): Uint8Array | string;
      getHash_asU8(): Uint8Array;
      getHash_asB64(): string;
      setHash(value: Uint8Array | string): Outpoint;

      getIndex(): number;
      setIndex(value: number): Outpoint;

      serializeBinary(): Uint8Array;
      toObject(includeInstance?: boolean): Outpoint.AsObject;
      static toObject(includeInstance: boolean, msg: Outpoint): Outpoint.AsObject;
      static serializeBinaryToWriter(message: Outpoint, writer: jspb.BinaryWriter): void;
      static deserializeBinary(bytes: Uint8Array): Outpoint;
      static deserializeBinaryFromReader(message: Outpoint, reader: jspb.BinaryReader): Outpoint;
    }

    export namespace Outpoint {
      export type AsObject = {
        hash: Uint8Array | string,
        index: number,
      }
    }

  }


  export class Output extends jspb.Message {
    getIndex(): number;
    setIndex(value: number): Output;

    getValue(): number;
    setValue(value: number): Output;

    getPubkeyScript(): Uint8Array | string;
    getPubkeyScript_asU8(): Uint8Array;
    getPubkeyScript_asB64(): string;
    setPubkeyScript(value: Uint8Array | string): Output;

    getAddress(): string;
    setAddress(value: string): Output;

    getScriptClass(): string;
    setScriptClass(value: string): Output;

    getDisassembledScript(): string;
    setDisassembledScript(value: string): Output;

    getSlpToken(): SlpToken | undefined;
    setSlpToken(value?: SlpToken): Output;
    hasSlpToken(): boolean;
    clearSlpToken(): Output;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Output.AsObject;
    static toObject(includeInstance: boolean, msg: Output): Output.AsObject;
    static serializeBinaryToWriter(message: Output, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Output;
    static deserializeBinaryFromReader(message: Output, reader: jspb.BinaryReader): Output;
  }

  export namespace Output {
    export type AsObject = {
      index: number,
      value: number,
      pubkeyScript: Uint8Array | string,
      address: string,
      scriptClass: string,
      disassembledScript: string,
      slpToken?: SlpToken.AsObject,
    }
  }

}

export class MempoolTransaction extends jspb.Message {
  getTransaction(): Transaction | undefined;
  setTransaction(value?: Transaction): MempoolTransaction;
  hasTransaction(): boolean;
  clearTransaction(): MempoolTransaction;

  getAddedTime(): number;
  setAddedTime(value: number): MempoolTransaction;

  getAddedHeight(): number;
  setAddedHeight(value: number): MempoolTransaction;

  getFee(): number;
  setFee(value: number): MempoolTransaction;

  getFeePerKb(): number;
  setFeePerKb(value: number): MempoolTransaction;

  getStartingPriority(): number;
  setStartingPriority(value: number): MempoolTransaction;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MempoolTransaction.AsObject;
  static toObject(includeInstance: boolean, msg: MempoolTransaction): MempoolTransaction.AsObject;
  static serializeBinaryToWriter(message: MempoolTransaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MempoolTransaction;
  static deserializeBinaryFromReader(message: MempoolTransaction, reader: jspb.BinaryReader): MempoolTransaction;
}

export namespace MempoolTransaction {
  export type AsObject = {
    transaction?: Transaction.AsObject,
    addedTime: number,
    addedHeight: number,
    fee: number,
    feePerKb: number,
    startingPriority: number,
  }
}

export class UnspentOutput extends jspb.Message {
  getOutpoint(): Transaction.Input.Outpoint | undefined;
  setOutpoint(value?: Transaction.Input.Outpoint): UnspentOutput;
  hasOutpoint(): boolean;
  clearOutpoint(): UnspentOutput;

  getPubkeyScript(): Uint8Array | string;
  getPubkeyScript_asU8(): Uint8Array;
  getPubkeyScript_asB64(): string;
  setPubkeyScript(value: Uint8Array | string): UnspentOutput;

  getValue(): number;
  setValue(value: number): UnspentOutput;

  getIsCoinbase(): boolean;
  setIsCoinbase(value: boolean): UnspentOutput;

  getBlockHeight(): number;
  setBlockHeight(value: number): UnspentOutput;

  getSlpToken(): SlpToken | undefined;
  setSlpToken(value?: SlpToken): UnspentOutput;
  hasSlpToken(): boolean;
  clearSlpToken(): UnspentOutput;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): UnspentOutput.AsObject;
  static toObject(includeInstance: boolean, msg: UnspentOutput): UnspentOutput.AsObject;
  static serializeBinaryToWriter(message: UnspentOutput, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): UnspentOutput;
  static deserializeBinaryFromReader(message: UnspentOutput, reader: jspb.BinaryReader): UnspentOutput;
}

export namespace UnspentOutput {
  export type AsObject = {
    outpoint?: Transaction.Input.Outpoint.AsObject,
    pubkeyScript: Uint8Array | string,
    value: number,
    isCoinbase: boolean,
    blockHeight: number,
    slpToken?: SlpToken.AsObject,
  }
}
// ...
export namespace SlpToken {
  export type AsObject = {
    tokenId: Uint8Array | string,
    amount: string,
    isMintBaton: boolean,
    address: string,
    decimals: number,
    slpAction: SlpAction,
    tokenType: SlpTokenType,
  }
}
// ...
export namespace SlpTransactionInfo {
  export type AsObject = {
    slpAction: SlpAction,
    validityJudgement: SlpTransactionInfo.ValidityJudgement,
    parseError: string,
    tokenId: Uint8Array | string,
    burnFlagsList: Array<SlpTransactionInfo.BurnFlags>,
    v1Genesis?: SlpV1GenesisMetadata.AsObject,
    v1Mint?: SlpV1MintMetadata.AsObject,
    v1Send?: SlpV1SendMetadata.AsObject,
    v1Nft1ChildGenesis?: SlpV1Nft1ChildGenesisMetadata.AsObject,
    v1Nft1ChildSend?: SlpV1Nft1ChildSendMetadata.AsObject,
  }

  export enum ValidityJudgement { 
    UNKNOWN_OR_INVALID = 0,
    VALID = 1,
  }

  export enum BurnFlags { 
    BURNED_INPUTS_OUTPUTS_TOO_HIGH = 0,
    BURNED_INPUTS_BAD_OPRETURN = 1,
    BURNED_INPUTS_OTHER_TOKEN = 2,
    BURNED_OUTPUTS_MISSING_BCH_VOUT = 3,
    BURNED_INPUTS_GREATER_THAN_OUTPUTS = 4,
  }

  export enum TxMetadataCase { 
    TX_METADATA_NOT_SET = 0,
    V1_GENESIS = 6,
    V1_MINT = 7,
    V1_SEND = 8,
    V1_NFT1_CHILD_GENESIS = 9,
    V1_NFT1_CHILD_SEND = 10,
  }
}
// ...
export class SlpTokenMetadata extends jspb.Message {
  getTokenId(): Uint8Array | string;
  getTokenId_asU8(): Uint8Array;
  getTokenId_asB64(): string;
  setTokenId(value: Uint8Array | string): SlpTokenMetadata;

  getTokenType(): SlpTokenType;
  setTokenType(value: SlpTokenType): SlpTokenMetadata;

  getV1Fungible(): SlpTokenMetadata.V1Fungible | undefined;
  setV1Fungible(value?: SlpTokenMetadata.V1Fungible): SlpTokenMetadata;
  hasV1Fungible(): boolean;
  clearV1Fungible(): SlpTokenMetadata;

  getV1Nft1Group(): SlpTokenMetadata.V1NFT1Group | undefined;
  setV1Nft1Group(value?: SlpTokenMetadata.V1NFT1Group): SlpTokenMetadata;
  hasV1Nft1Group(): boolean;
  clearV1Nft1Group(): SlpTokenMetadata;

  getV1Nft1Child(): SlpTokenMetadata.V1NFT1Child | undefined;
  setV1Nft1Child(value?: SlpTokenMetadata.V1NFT1Child): SlpTokenMetadata;
  hasV1Nft1Child(): boolean;
  clearV1Nft1Child(): SlpTokenMetadata;

  getTypeMetadataCase(): SlpTokenMetadata.TypeMetadataCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): SlpTokenMetadata.AsObject;
  static toObject(includeInstance: boolean, msg: SlpTokenMetadata): SlpTokenMetadata.AsObject;
  static serializeBinaryToWriter(message: SlpTokenMetadata, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): SlpTokenMetadata;
  static deserializeBinaryFromReader(message: SlpTokenMetadata, reader: jspb.BinaryReader): SlpTokenMetadata;
}
// ...
export enum SlpTokenType { 
  VERSION_NOT_SET = 0,
  V1_FUNGIBLE = 1,
  V1_NFT1_CHILD = 65,
  V1_NFT1_GROUP = 129,
}